requirements.txt
Flask
Flask-SQLAlchemy
pymysql
flask-cors
cryptography
gunicorn
python-dotenv
boto3
sentence-transformers>=2.3.0
nltk>=3.8.1
bertopic>=0.16.0
scikit-learn>=1.3.0
pandas>=2.0.0
beautifulsoup4>=4.12.0
geopy>=2.3.0
requests>=2.31.0
pypdf>=3.0.0
python-docx>=0.8.11
APScheduler>=3.10.0